requirements.txt
python-dwca-reader
pandas
asyncio
aiohttp
duckdb
requests
tqdm
google-cloud-bigquery
db-dtypes
validators
python-dotenv
dlt
